This update:
1. Adds tornado genesis parameters to all NMM cores.  This is turned
on or off by the new tc_option namelist variable, which is off by
default.

2. Corrects a bug in reg_parse.c that could result in incorrect code
for any NMM core: the mp_var and nmm_v_grid variables were not
initialized while parsing the registry.

3. Modifies the module_wrf_error to use format(A) while printing.
This works around a bug in the Intel compiler wherein all lines are
wrapped at 80 characters, causing problems for anything that tries to
parse the rsl files.




git-svn-id: https://svn-wrf-model.cgd.ucar.edu/branches/hwrf-tc-diag-2@7093 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d